utils: switch from pwclient to git-pw
Easier not to mix the different tools but also because pwclient seems no
longer returning what is requested.

pwclient was mainly used because we don't need to specify all states we
want: by default, it lists '*' which is not possible with git-pw, we
need to use '--state x' multiple times. But anyway, we always have to
filter the state.

Also a good thing with 'git-pw' is that it can search for patches with a
certain name.

While at it, we also grab more details when listing all patches: series,
url, submitter

We can also remove no longer used scripts and merge
patch-subject-get-ids.sh in patch-subject-accept.sh as it is only used
there.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>