mptcp: never re-inject data on a single subflow
If the MPTCP-level data ack is delayed WRT to the TCP-level ack,
there is a single subflow and the user-space stops pushing data,
the MPTCP-level retransmit may trigger in, fooling (disallowing)
infinite mapping fallback.

All the above is triggered quite reliably by the self-tests, once
we move the MPTCP receive path under the msk socket lock - delaying
the MPTCP-level acks.

Plain TCP is good enough to handle retransmissions as needed.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Message-Id: <cfe4fd464d2e7701008b5e1b39030b74f3e9a606.1661339834.git.pabeni@redhat.com>