selftests: mptcp: set test-case parameters consistently
The mptcp_join test-cases depend on several parameters. The caller
provide most of them as command line arguments for the main test
function. That makes the test-cases are do read.

Additionally we sometimes need to add more parameters, which currently
is currently error prone and makes the test-cases even more hard to
follow.

Refactor the infrastructure to pass test-case parameters as enviromental
variables, and update all the existing tests accordingly.

Suggested-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Message-Id: <2abed554f2050c7626dede74c92a181506a92682.1665779933.git.pabeni@redhat.com>