cirrus: save ccache dir on failure
When switching to a new kernel version, the cache often cannot be
re-used. It means the compilation takes more time and we can hit the
global timeout...

Force uploading the cache even in case of failure (if it is supported),
at least to have the next build benefiting from the updated build cache.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>