cirrus: force compilation only
The recent switch to a new kernel version makes the CI taking a longer
time to compile everything.

We end up hitting the max timeout for the debug mode and the updated
ccache directory is not updated. So the next builds in debug mode does
that again...

Just force here one build just doing the compilation bit and starting
the VM, that's it, just to fill cache and upload them.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>