Squash-to (part 2): "mptcp: add subflow_v(4,6)_send_synack()"
mptcp: move mptcp_set_owner_r() signature with others from protocol.c

Because this function is not specific to fastopen functions.

I added it at the end of the list there because there is no real order.
Maybe we should order them differently to reduce conflicts in case of
backports? Anyway, that would not be don't in this patch.

Link: https://lore.kernel.org/r/20221114164601.861295-4-matthieu.baerts@tessares.net
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>