ci: validate mptcp_join selftest in //
The goal is to have 2x2 builds now:
 - with and without mptcp_join selftest
 - with and without debug kconfig

We should then have 4 results to parse.

Hopefully, this should accelerate builds.

For the moment, mptcp_join selftest takes ~half of the time, e.g. from a
recent debug build:

	clone:		12.17.39	4m
	preparation:	12.22.42	5m (ccache was up to date)
	init:				9m
	kunit:		12.22.44	2s
	diag:		12.23.34	50s
	connect:	12.34.36	11m
	join:		13.21.14	43m
	sockopt:	13.22.26	72s
	pm_netlink:	13.25.19	3m
	simult_flow:	13.26.54	95s
	userspace_pm:	13.28.35	101s
	connect mmap:	13.38.59	10m
	selftests:			1h16
	add_addr:	13.39.59	1m
	dss:		13.40.43	44s
	fastclose:	13.41.14	31s
	fastopen:	13.42.09	55s
	mp_capable:	13.43.22	73s
	mp_join:	13.43.53	31s
	mp_prio:	13.44.08	15s
	mp_reset:	13.44.32	24s
	regressions:	13.44.55	23s
	sockopts:	13.45.20	25s
	syscalls:	13.46.24	64s
	packetdrill:			7m25

So if we extract mptcp join selftest to a dedicated build, that could
help having results faster.

It is not clear if we can have two 'matrix' in Cirrus' 'env' section.

Suggested-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>