cirrus: increase the allocated memory
Builds with a debug kernel config have been failing since the merge with
Linus tree in preparation of the future v6.2.

In the logs, all we could see is:

  [14:13:48.867]   AR      built-in.a
  [14:13:49.221]   AR      vmlinux.a
  [14:13:52.228]   LD      vmlinux.o
  [14:14:04.721] Killed
  [14:14:04.721] make[2]: *** [scripts/Makefile.vmlinux_o:61: vmlinux.o] Error 137

From: https://cirrus-ci.com/task/5046665359392768

Not very helpful.

The graphs showing the RAM usage were always suggesting memory usage was
increasing but approximatively using around max half of the 4GB.

Tests have shown that when increasing the RAM allocated to the test
machine, the build can now pass. Let's do that, Cirrus allows us to take
more anyway.

Note that some work is being done on the kernel side to reduce the
maximum memory usage:

  https://lore.kernel.org/lkml/20221216-objtool-memory-v2-0-17968f85a464@weissschuh.net/

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>