mptcp: avoid unneeded __mptcp_nmpc_socket() usage
In a few spots the mptcp code invokes the __mptcp_nmpc_socket() helper
multiple times under the same socket lock scope. Additionally, in such
places, the socket status ensure that threre is not an MP capable
handshake running.

Under the above condition we can replace the later __mptcp_nmpc_socket()
helper invocation with direct access to the msk->subflow pointer and
better document such access is not supposed to fail with WARN().

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Message-Id: <67ebfe6c67672d72728d02bb92503843aaae826e.1673629755.git.pabeni@redhat.com>