mptcp: do not propagate fallback subflow status on error
Currently the subflow error report callback unconditionally
propagates the fallback subflow status to the owning msk.

That is not needed as there is a great deal of infrastructure
trying to propagate correctly the fallback subflow status to
the owning mptcp socket, e.g. via mptcp_subflow_eof() and
subflow_sched_work_if_closed().

And in some circumstances - specifically if the msk is already
orphaned - it prevents the code from correctly tracking the msk
moving to the TCP_CLOSE state and doing the appropriate
cleanup.

All the above causes increasing memory usage over time and
sporadic self-tests failures.

Address the issue simply removing the unneeded state update.

Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/339
Fixes: 15cc10453398 ("mptcp: deliver ssk errors to msk")
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Reviewed-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <97f7deb4132d567132088d43f3e7f5ff0602a45b.1674668294.git.pabeni@redhat.com>