Squash-to: DO-NOT-MERGE: mptcp: use kmalloc on kasan build
This is an extra check mainly for the CIs: to make sure we don't call
tcp_done() on an already closed socket as it happened in the past. If we
do such call, better to catch the error earlier.

This is not really related to the Squash-to patch (mptcp: use kmalloc on
kasan build) but it is just not to create yet another specific commit
for the CI added to both export and export-net branches. This commit can
also be renamed to something like:

  DO-NOT-MERGE: mptcp: improve code coverage for CI

WDYT?

Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/362
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <20230221-mptcp-warn-tcp-done-close-v1-2-d248242b29b6@tessares.net>