DO-NOT-MERGE: git markup: fixes other trees
All commits older than this one are fixes for other trees than the -net
one or imported from elsewhere. We should not send these commits to
either -net or net-next or at least not as "MPTCP fixes".

Typically they are fixing issues blocking our tests or dev.

Following commits are MPTCP-related fixes for the -net tree.

This commit is useful to easily find where are the fixes from/for other
trees.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>