gh: build: also use cache for patches in review
With patches in review, 'github.ref_name' will be something like
'patchew/<message_id>'. Because of that, we would not be able to find
any cache with this name.

Instead, we should use cache linked to 'export' or 'export-net' branch.
That's what we do here by checking if a parent commit is the one on top
of 'export-net'. If not, we default on 'export'.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>