Squash to "selftests: mptcp: add MPTCP_FULL_INFO testcase"
Relax the expected returned sizes from a getsockopt(MPTCP_FULL_INFO):
the sizes should be positive and the user one should be the min between
the one given by the userspace and the one from the kernelspace.

By doing that, we support new fields added on the userspace or the
kernelspace sides.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <20230406-mptcp-issue-368-selftests-old-kernels-v3-32-a1f05258292b@tessares.net>