net: skip printing "link become ready" v6 msg
This following message is printed in the console each time a network
device configured with an IPv6 addresses is ready to be used:

  ADDRCONF(NETDEV_CHANGE): <iface>: link becomes ready

When netns are being extensively used -- e.g. by re-creating netns with
veth to discuss with each other for testing purposes like mptcp_join.sh
selftest does -- it generates a lot of messages: more than 700 when
executing mptcp_join.sh with the latest version.

=========
== RFC ==
=========

TL;DR: can we move this message to the debug level? Or is it better with
a sysctl knob? Or something else?

When looking at commit 3c21edbd1137 ("[IPV6]: Defer IPv6 device
initialization until the link becomes ready.") which introduces this new
message, it seems it had been added to verify that the new feature was
working as expected. It could have then used a lower level than "info".

It is unclear if this message can be useful. Maybe it can be used as a
sign to know if there is something wrong, e.g. if a device is being
regularly reconfigured by accident? But even then, I don't think that
was its goal at the first place and clearly there are better ways to
monitor and diagnose such issues. Do you see any usages?

If this message is not that useful, it is probably better to simply
lower its level, similar to commit 7c62b8dd5ca8 ("net/ipv6: lower the
level of "link is not ready" messages"). If we can take this direction,
we will just need to switch from pr_info() to pr_debug().

If this message can be useful in many situations, it would be good to
have a way to turn it off because in some other situations, it floods
the logs without providing any useful input. The proposition here is to
have a new per netns sysctl knob to easily skip this specific message
when needed. If we prefer to take this direction, we will still need to
document the new knob and the modification in the MPTCP selftest should
be done in a separated commit.

Adding a new sysctl entry just for that seems a bit "heavy", maybe there
are better ways that are still easy to put in place?

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <20230601-net-next-skip_print_link_becomes_ready-v1-1-c13e64c14095@tessares.net>